+++ src/plugins/apollo/apolloClient.ts
@@ -1,7 +1,7 @@
 import { ApolloClient, InMemoryCache } from '@apollo/client';
 
 const client = new ApolloClient({
-  uri: process.env.API_HOST_URL || 'https://jabref-online.herokuapp.com',
+  uri: process.env?.API_HOST_URL || 'https://jabref-online.herokuapp.com',
   cache: new InMemoryCache(),
 });
 export default client;
