+++ src/taskpane/routes/Index.tsx
@@ -1,8 +1,6 @@
-/* eslint-disable import/extensions */
-/* eslint-disable import/no-unresolved */
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
-import Context from '../contexts';
+import Context from '../contexts/Index';
 import Layout from '../layout/Layout';
 import Auth from '../pages/Auth';
 import ProtectedRoutes from './ProtectedRoutes';
